create hub client
