+++ tests/unit/sagemaker/jumpstart/curated_hub/test_curated_hub_full_workflow.py
@@ -9,7 +9,7 @@ from botocore.client import ClientError
 
 
 from sagemaker.jumpstart.curated_hub.jumpstart_curated_public_hub import JumpStartCuratedPublicHub
-from sagemaker.jumpstart.curated_hub.jumpstart_curated_public_hub import PublicModelId
+from sagemaker.jumpstart.curated_hub.utils import PublicModelId
 
 
 class JumpStartCuratedPublicHubTest(unittest.TestCase):
